[iOS] Chore: 뷰객체 들을 각각의 file을 만들어 옮김
