+++ iOS/Sidedish/Sidedish/ViewControllers/DetailViewController.swift
@@ -82,121 +82,3 @@ final class DetailViewController: UIViewController {
         }
     }
 }
-
-final class DetailVerticalScrollView: UIScrollView {
-    private let thumbScrollView = ThumbScrollView()
-    private let detailStackView = DetailStackView()
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        translatesAutoresizingMaskIntoConstraints = false
-        configureThumbScrollView()
-        configureDetailStackView()
-    }
-    
-    required init?(coder: NSCoder) {
-        super.init(coder: coder)
-        translatesAutoresizingMaskIntoConstraints = false
-        configureThumbScrollView()
-        configureDetailStackView()
-    }
-    
-    private func configureThumbScrollView() {
-        addSubview(thumbScrollView)
-        
-        thumbScrollView.widthAnchor.constraint(equalTo: widthAnchor).isActive = true
-        thumbScrollView.topAnchor.constraint(equalTo: topAnchor).isActive = true
-        thumbScrollView.leadingAnchor.constraint(equalTo: leadingAnchor).isActive = true
-        contentLayoutGuide.topAnchor.constraint(equalTo: thumbScrollView.topAnchor).isActive = true
-    }
-    
-    private func configureDetailStackView() {
-        addSubview(detailStackView)
-        
-        detailStackView.widthAnchor.constraint(equalTo: widthAnchor).isActive = true
-        detailStackView.topAnchor.constraint(equalTo: thumbScrollView.bottomAnchor).isActive = true
-        detailStackView.leadingAnchor.constraint(equalTo: leadingAnchor).isActive = true
-        contentLayoutGuide.bottomAnchor.constraint(equalTo: detailStackView.bottomAnchor).isActive = true
-    }
-    
-    func addThumb(imageView: UIImageView){
-        thumbScrollView.addToStack(view: imageView)
-    }
-    
-    func addDetail(imageView: UIImageView) {
-        let ratio =  imageView.frame.height / imageView.frame.width
-        
-        detailStackView.addArrangedSubview(imageView)
-        imageView.widthAnchor.constraint(equalTo: detailStackView.widthAnchor).isActive = true
-        imageView.heightAnchor.constraint(equalTo: imageView.widthAnchor, multiplier: ratio).isActive = true
-    }
-}
-
-final class ThumbScrollView: UIScrollView {
-    private let thumbStackView = ThumbStackView()
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        translatesAutoresizingMaskIntoConstraints = false
-        isPagingEnabled = true
-        configureThumbStackView()
-    }
-    
-    required init?(coder: NSCoder) {
-        super.init(coder: coder)
-        translatesAutoresizingMaskIntoConstraints = false
-        isPagingEnabled = true
-        configureThumbStackView()
-    }
-    
-    private func configureThumbStackView() {
-        addSubview(thumbStackView)
-        
-        thumbStackView.heightAnchor.constraint(equalTo: heightAnchor).isActive = true
-        thumbStackView.topAnchor.constraint(equalTo: topAnchor).isActive = true
-        thumbStackView.leadingAnchor.constraint(equalTo: leadingAnchor).isActive = true
-        contentLayoutGuide.widthAnchor.constraint(equalTo: thumbStackView.widthAnchor).isActive = true
-    }
-    
-    func addToStack(view: UIView) {
-        thumbStackView.addArrangedSubview(view)
-        view.widthAnchor.constraint(equalTo: widthAnchor).isActive = true
-        view.heightAnchor.constraint(equalTo: view.widthAnchor).isActive = true
-    }
-}
-
-final class ThumbStackView: UIStackView {
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        configure()
-    }
-    
-    required init(coder: NSCoder) {
-        super.init(coder: coder)
-        configure()
-    }
-    
-    private func configure() {
-        translatesAutoresizingMaskIntoConstraints = false
-        axis = .horizontal
-        distribution = .fill
-    }
-}
-
-final class DetailStackView: UIStackView {
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        configureView()
-    }
-    
-    required init(coder: NSCoder) {
-        super.init(coder: coder)
-        configureView()
-    }
-    
-    private func configureView() {
-        translatesAutoresizingMaskIntoConstraints = false
-        axis = .vertical
-        distribution = .fill
-    }
-}
