+++ iOS/Sidedish/Sidedish/Views/Detail's/DetailStackView.swift
@@ -1,0 +1,27 @@
+//
+//  DetailStackView.swift
+//  Sidedish
+//
+//  Created by kimdo2297 on 2020/05/02.
+//  Copyright © 2020 Jason. All rights reserved.
+//
+
+import UIKit
+
+final class DetailStackView: UIStackView {
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        configureView()
+    }
+    
+    required init(coder: NSCoder) {
+        super.init(coder: coder)
+        configureView()
+    }
+    
+    private func configureView() {
+        translatesAutoresizingMaskIntoConstraints = false
+        axis = .vertical
+        distribution = .fill
+    }
+}
