+++ iOS/Sidedish/Sidedish/Views/Detail's/ThumbStackView.swift
@@ -1,0 +1,27 @@
+//
+//  ThumbStackView.swift
+//  Sidedish
+//
+//  Created by kimdo2297 on 2020/05/02.
+//  Copyright © 2020 Jason. All rights reserved.
+//
+
+import UIKit
+
+final class ThumbStackView: UIStackView {
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        configure()
+    }
+    
+    required init(coder: NSCoder) {
+        super.init(coder: coder)
+        configure()
+    }
+    
+    private func configure() {
+        translatesAutoresizingMaskIntoConstraints = false
+        axis = .horizontal
+        distribution = .fill
+    }
+}
